requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.10.15
pymongo==4.7.2
python-dotenv==1.0.1
